Move tabData tests to own file.
