+++ src/tests/index.html
@@ -64,7 +64,8 @@
     <script src="tests/htmlutils.js"></script>
     <script src="tests/multiDomainFirstParties.js"></script>
     <script src="tests/storage.js"></script>
-    <script src="tests/baseDomain.js"></script>
+    <script src="tests/tabData.js"></script>
+    <!-- URL/host tools: --><script src="tests/baseDomain.js"></script>
     <script src="tests/utils.js"></script>
     <script src="tests/yellowlist.js"></script>
   </body>
